contest page updates
