+++ client/src/components/ContestComponents/CountdownTimer/CountdownTimer.css
@@ -1,0 +1,10 @@
+.timer{
+    background: rgb(0,0,0, 0.9);
+    color: white;
+    font-size: 22px;
+    display: grid;
+    grid-template-columns: repeat(7, 1fr);
+    text-align: center;
+    border-radius: 3px;
+    padding: 0rem 1rem;
+}
