+++ client/src/screens/ContestScreens/ContestPage/ContestPage.css
@@ -6,17 +6,6 @@ body{
     font-size: 1.2rem;
 }
 
-.timer{
-    background: rgb(0,0,0, 0.9);
-    color: white;
-    font-size: 22px;
-    display: grid;
-    grid-template-columns: repeat(7, 1fr);
-    text-align: center;
-    border-radius: 3px;
-    padding: 0rem 2rem;
-}
-
 .timer-container h5, 
 .create-submission h5 {
   font-size: 1.5rem;
