+++ client/src/screens/ContestScreens/ContestPage/ContestPage.jsx
@@ -1,5 +1,4 @@
-import React, { Fragment, useState, useRef, useEffect } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
+import React from 'react';
 import Search from '../../../components/Form/Search';
 import { useStateValue } from "../../../providers/CurrentUserProvider";
 import AccountCircleIcon from "@material-ui/icons/AccountCircle";
@@ -8,47 +7,12 @@ import Button from '@material-ui/core/Button';
 import Layout from '../../../layout/Layout'
 import TextField from '@material-ui/core/TextField';
 import AddCircle from '@material-ui/icons/AddCircle';
+import CountdownTimer from '../../../components/ContestComponents/CountdownTimer/CountdownTimer';
+import ContestChat from '../../../components/ContestComponents/ContestChat/ContestChat';
 
 
 function ContestPage() {
     const [{ currentUser }, dispatch] = useStateValue();
-    const [timerDays, setTimerDays] = useState('00');
-    const [timerHours, setTimerHours] = useState('00');
-    const [timerMinutes, setTimerMinutes] = useState('00');
-    const [timerSeconds, setTimerSeconds] = useState('00');
-
-    let interval = useRef()
-
-    const startTimer = () => {
-        const countdownDate = new Date('May 30, 2020 00:00:00').getTime();
-
-        interval = setInterval(() => {
-            const now = new Date().getTime();
-            const difference = countdownDate - now;
-
-            const days = Math.floor(difference / (1000 * 60 *60 * 24));
-            const hours = Math.floor((difference % (1000 * 60 *60 * 24) / (1000 * 60 * 60)));
-            const minutes = Math.floor((difference % (1000 * 60 * 60)) / (1000 * 60));
-            const seconds = Math.floor((difference % (1000 * 60)) / 1000);
-
-            if (difference < 0) {
-                // stop timer
-                clearInterval(interval.current)
-            } else {
-                setTimerDays(days);
-                setTimerHours(hours);
-                setTimerMinutes(minutes);
-                setTimerSeconds(seconds);
-            }
-        }, 1000)
-    }
-
-    useEffect(() => {
-      startTimer();  
-        return () => {
-        clearInterval(interval.current);
-        }
-    })
 
     return (
         <Layout>
@@ -60,27 +24,7 @@ function ContestPage() {
         <main>
             <section className='timer-container'>
                 <h5>Contest Ends In:</h5>
-            <div className='timer'>
-                <section>
-                    <p>{timerDays}</p>
-                    <p><small>Days</small></p>
-                </section>
-                <p>:</p>
-                <section>
-                    <p>{timerHours}</p>
-                    <p><small>Hours</small></p>
-                </section>
-                <p>:</p>
-                <section>
-                    <p>{timerMinutes}</p>
-                    <p><small>Minutes</small></p>
-                </section>
-                <p>:</p>
-                <section>
-                    <p>{timerSeconds}</p>
-                    <p><small>Seconds</small></p>
-                </section>
-            </div>
+                <CountdownTimer />
             <h5>Contest Rules</h5>
             <div>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit.<br/> 
@@ -117,6 +61,7 @@ function ContestPage() {
             </div>
         </main>
         <hr style={{ margin: "0rem 2rem"}}/>
+        <ContestChat />
         <section>
         <div className='submission-name'>View Contest Entries</div>
         </section>
